fix: Remove flaky test from suite
test_get_preview was comparing a downloaded image to a non-static image downloaded from the LDS.
This results in a flaky test and having to update the image in the repo often.
This is not good for a stable test suite and has resulted in failing CI builds in the repository.
Removing the flaky test here.